Revision of the OutbreakCode project. Previous version now under OutbreakCodeJan2012. Most revisions made in the reader package concerning the infection tree and outbreaks.
IMP: Added a new package lineagetree. The LineageClass in this package takes in the infection tree as an input and transforms it to a lineage/transmission tree (a binary tree). 

Coming next: Methods to calculate MRCA, tree balance statistics and sampling from the transmission/lineage tree. 

Packages 'reader' and 'lineagetree' are most likely to be updated following the above upcoming additions. Other classes may also be modified (although minor changes expected).
--
sja